+++ src/utils/formDataGetter.ts
@@ -1,5 +1,7 @@
 export const getFormData = (target: EventTarget) => {
   if (!(target instanceof HTMLFormElement)) return;
 
-  return new FormData(target);
+  const formData = new FormData(target);
+
+  return Object.fromEntries(formData);
 };
